Create Cart
Allowing users to select quantities of products they want to buy.
By submitting the form, a new Cart object is created and associated to the user's lastCart.
If a lastCart already existed, it will be removed and replaced by the new one.
Last, but not least, the newly created cart can only be seen/modified by the user who created it.