Purchases
Allowing users to finalize their purchases by choosing a payment method.
A purchase contains a payment method plus the cart.
The cart will also have a link to the purchase.
The cart will no longer be the user's last cart, and the purchase object will be added as a "purchase relation" of the user.
Users can view carts they've purchased. The view will show the payment date and purchase method.